# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Reservation System API",
        "description": "API REST para el sistema de reservas de hotel. Incluye gestión de amenities, tipos de habitación y habitaciones físicas.",
        "contact": {
            "name": "API Support",
            "email": "adiazy@uamv.edu.ni"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener lista de reservas",
                "description": "Lista de reservas\n\nFiltros opcionales vía query params:\n- ?status=CONFIRMED\n- ?from=2025-01-01&to=2025-01-31 (sobre check_in)\n- ?guest_id=5",
                "operationId": "f10e68ff0ce263ef754b2f640eda8d93",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDING",
                                "CONFIRMED",
                                "CHECKED_IN",
                                "CHECKED_OUT",
                                "CANCELLED",
                                "NO_SHOW"
                            ],
                            "example": "CONFIRMED"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha desde (formato: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha hasta (formato: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-31"
                        }
                    },
                    {
                        "name": "guest_id",
                        "in": "query",
                        "description": "Filtrar por ID de huésped",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reservas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reserva",
                "description": "Crear una nueva reserva\n\nCrea la reserva y sus reservation_rooms en una transacción.\nSi rooms viene en el payload, crea los registros en reservation_rooms.\nOpcionalmente recalcula total_amount como suma de noches × precio.",
                "operationId": "4f59911f11563f79b4f01bd02222607d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "guest_id",
                                    "check_in",
                                    "check_out",
                                    "adults"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "ABC123"
                                    },
                                    "guest_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "PENDING",
                                            "CONFIRMED",
                                            "CHECKED_IN",
                                            "CHECKED_OUT",
                                            "CANCELLED",
                                            "NO_SHOW"
                                        ],
                                        "example": "CONFIRMED",
                                        "nullable": true
                                    },
                                    "channel": {
                                        "type": "string",
                                        "enum": [
                                            "DIRECT",
                                            "PHONE",
                                            "WALKIN",
                                            "OTA"
                                        ],
                                        "example": "DIRECT",
                                        "nullable": true
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-10"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-12"
                                    },
                                    "adults": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "children": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "maxLength": 3,
                                        "example": "NIO",
                                        "nullable": true
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 160,
                                        "nullable": true
                                    },
                                    "paid_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50,
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Llegan tarde.",
                                        "nullable": true
                                    },
                                    "rooms": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "room_type_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "room_id": {
                                                    "type": "integer",
                                                    "example": 5,
                                                    "nullable": true
                                                },
                                                "rate_plan_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "nightly_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 80
                                                },
                                                "date_from": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-01-10"
                                                },
                                                "date_to": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-01-12"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
    },
...
}